third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylehash-allowed-expected.txt

ALERT: PASS: The 'p' element's text is green, which means the style was correctly applied.
This tests the result of a valid style hash. It passes if this text is green, and a "PASS" alert is fired.
